refactor: Move publish logic to lib/publishPackages.ts and call directly
Extract the tarball publishing logic into an importable function in
lib/publishPackages.ts and call it directly from registrySetup instead
of spawning a subprocess. Remove the old publish-packages.ts script.
